# Performance improvements on queries