--- a/app.py
+++ b/app.py
@@ -80,6 +80,8 @@ app.layout = html.Div([
     dummy_div,
     dcc.Store(id='stored-data'),
     dcc.Store(id='stored-data-func'),
+    dcc.Store(id='stored-data-education-total'),
+    dcc.Store(id='stored-data-education-outcome')
 ])
 
 @app.callback(