--- a/pages/education.py
+++ b/pages/education.py
@@ -19,6 +19,8 @@ layout = html.Div(children=[
             html.Div(id='education-content'),
         ]),
     ),
+    dcc.Store(id='stored-data-education-total'),
+    dcc.Store(id='stored-data-education-outcome')
 ])