Tried the following: 

1. Async rendering - now the query should be mapped to a specific plot. This enables a specific plot to be rendered after only the relevant queries are completed.
2. Pushed down ORDER opearation after the filtering (now the filtering is done by Pandas)
3. Pre-execution of the queries. (the new tables are created in this PR. https://github.com/dime-worldbank/mega-boost/pull/17)

|                                            | Before (seconds) | After (seconds) | Methods                                              |
|--------------------------------------------|------------------|-----------------|------------------------------------------------------|
| get_expenditure_w_porverty_by_country_year | 2.871            | 0.651           | 1. SORT OPERATION PUSHDOWN  2. Pre-join in databricks |
| get_expenditure_by_country_func_econ_year  | 2.017            | 0.763           | 1. SORT OPERATION PUSHDOWN                           |
| get_expenditure_by_country_func_year       | 3.099e-06        | 0.00            | 1. SORT OPERATION PUSHDOWN                           |
| get_learning_poverty_rate                  | 1.668            | 0.713           | 1. SORT OPERATION PUSHDOWN                           |
| get_hd_index                               | 4.053e-06        | 5.198e-06       | 1. SORT OPERATION PUSHDOWN                           |