@weilu 2024-07-31
I'm surprised sorting in databricks is so much slower! Probably the overhead. Thank you for discovering this. Good to know.